Remove external HTTP routes
The Qserv Kafka bridge will (at least to start with) be driven only
by Kafka subscriptions. Remove the stub external route and its model,
leaving only the internal route that we'll use (for now) for health
checking. Drop the Slack reporting of exceptions for now; this will
probably be replaced with Sentry.