Add models of Kafka messages
Add Pydantic models for the basic Kafka messages for starting a job
and reporting job status, based on SQR-097. Change `qservID` to
`executionID`, add a `serialization` field next to the output format,
and don't return a URL to the results for right now until we figure
out how to handle the round trip. Serialize timestamps to milliseconds.