Add support for starting jobs
Add a storage backend for talking to the Qserv REST API and a service
layer that understands how to process job run requests. Start a query
on receipt of a Kafka message and immediately retrieve and report the
status of that query, which is published on the job status topic in
Kafka.