Update Kafka models for recent changes
Include the column type information in the `JobRun` model, which will
be used for output serialization. Rework how the output format is
represented to match recent changes in SQR-097. Include the
`resultLocation` attribute in the `JobRun` model so that it can be
included in the output for a completed job.