Add state polling and job status updates
Poll Qserv for state via the MySQL API, every second by default, and
look for any queries with status updates or that are missing from the
output (which will be the case if the query has finished). Get the
metadata for those queries and send Kafka messages recording the
updated status. This includes handling for aborted and completed jobs,
but the message for completed jobs does not yet include anything about
the result of the query.