Improve error handling of SQL failures
Somehow we saw an error from Qserv where the result table for a
completed query did not exist. When that happened, it caused the
Qserv Kafka bridge to repeatedly try to retrieve that row and did
not log the error properly because it was uncaught. Rewrite that
error into something we can catch and log properly and treat as a
failure.