Refactor query service layer
Previously, there was a lot of duplicate code between the query
monitor and the query service, since the query service also had to
process job status updates and generate an appropriate Kafka message.
Unify that code in the query service and have the query monitor use
the query service to process status updates.