Fix processing of requests after shutdown
When the application is shut down, the FastStream lifecycle hooks
shut down the broker built into the Kafka router. This means we
cannot use that broker in background jobs if we want uploading of
results to complete after the application starts shutting down.
Create a separate Kafka broker specifically for background jobs.

Most of this change is adding support for doing a proper test with
real Kafka started by testcontainers. Import the Kafka testcontainer
code from Safir, since it seems to work more reliably, and refactor
how the application is constructed so that it supports more
dependency injection of mocked brokers without using global variables,
since we have to defer Kafka setup until the testcontainer has been
started.

This approach is not very clean, and we may revisit this later when
we have a chance to think about a better pattern for Kafka testing.