Reduce garbage and copies in BINARY2 encoder
Simplify the BINARY2 encoder to use a smaller stack of async
generators and to encode directly into a `BytesIO` object (with an
intermediate `bytearray` per row). Batch base64-encoding as well
and only encode once we can generate `UPLOAD_BUFFER_SIZE` encoded
data (roughly), and use a `memoryview` on the underlying `BytesIO`
to avoid needing to make a copy.

Hopefully this will improve performance by decreasing the amount
of small garbage generated and reducing the number of memory copies.