Add metrics events to the Qserv Kafka bridge
Report metrics, separately for successful and failed queries. For
successful queries, report a variety of elapsed times and rates for
the portions of the query processing we may want to measure
separately.

In order to support total query execution time, add a start time to
the query state data. Add a dummy start time to existing state
entries that don't have one for backward compatibility. Now require
all state entries to have the flag indicating whether the query has
been dispatched to the backend worker.

Add more logging when we thought a query was complete but Qserv
reported that it was still executing.