Add memory leak test for encoder
Run through job submission and immediate completion with the Qserv
mock for 1,000 times and check the resulting memory consumption to
ensure we're not leaking too much memory. Python does quite a bit
of caching, so we need to allow for 300KB of memory growth, but that
number was stable between 1,000 and 5,000 iterations. Only run 1,000
iterations in the regular test since 5,000 takes a bit too long.