Format datetimes in the IVOA DALI format
In VOTables, database timestamp values should be formatted as strings
with an `xtype` of `timestamp`. Prior versions of the Qserv Kafka
bridge just called `str` on the Python `datetime` object returned by
SQLAlchemy, which is close but does not match the expected format
(space instead of `T` between the date and time, time given in
microsecond instead of millisecond precision, and with time zone
information appended).

The fully correct implementation would know whether a given output
field is supposed to be a timestamp and verify that the database
column was also a timestamp, but for the sake of expediency, avoid
requiring that information. Instead, whenever given a `datetime`
object to encode as a string, convert it to the IVOA DALI string
representation. (In other words, blindly assume that this is a
timestamp column.) This should work in all cases that arise in our
current database schema.

While adding tests for this case, also add an explicit test for
encoding a string that is variable length with an upper bound. Check
that the string is correctly truncated but that the string length is
prepended since the string counts as variable length for VOTable
encoding.