Improve testing with flaky Qserv mock
Change the default parametrization of the Qserv mock to only run in
normal mode, and instead parametrize for flaky mode only in the
service tests. This should still cover the same error handling, but
avoids adding delays to delay-sensitive tests via the Kafka handler
that depend on background timing delays.