Log metrics events for Qserv retries
Add a new metrics event for a successful Qserv API call after a
retry (failures will produce job errors and be accounted for
separately). Teach the retry decorator how to publish that metrics
event.

The code here is kind of ugly and hard to follow because Python
doesn't offer a great way of running code only if a block didn't
throw an exception (`try` can't be used without `except` or
`finally`).