Support dropping API version in Qserv REST calls
Currently, Qserv's REST API only supports a single API version at
a time. This means that if the Qserv Kafka bridge always sends an
API version, both components have to be upgraded in lockstep. Add
an option to disable sending the API version, which means Qserv will
make a best effort to interpret the call, with some potential errors
and misinterpretation if the semantics have changed.