Fix timings in query status
Fix query start and end times in status responses. Previously, the
start time and, for canceled or completed queries, the end time were
taken directly from the Qserv query information. This includes only
time spent by Qserv, which results in an inaccurate picture of the
query execution time in the TAP UWS table. Instead measure query
timings from the time the query was received by the bridge to when
it completes uploading of the results, or is aware of a failure.