Add support for TAP query quotas
Add a caching Gafaelfawr client and associated configuration to
retrieve the user's TAP quota for the configured TAP service name,
if it exists. Add tracking of the number of concurrent queries each
user has in progress. Using that information, impose a limit on
concurrent queries if the user's TAP quota indicates there should
be one, rejecting additional queries with an error until some other
query completes.

Add a background reconcile job that updates the counts of running
queries for each user based on a snapshot of other data stored in
Redis. This works from a snapshot without a concurrency lock, so it
may set the counts incorrectly, but it ensures that the system will
self-heal against any problem (server crashes, for instance) that
cause a failure to decrement the number of running queries. Try to
avoid strangeness from race conditions by only correcting negative
running query counts or counts where the Redis information appears
to be too high.