Tweak timeout and shield result shutdown
Increase the grace period between when we try to kill result upload
that is taking too long and when arq kills the worker, since we saw
one case of a failure to shut down cleanly before arq cancelled the
worker. Shield the close of the result generator that is reading from
MySQL so that even if we lose that raise, hopefully the SQL shutdown
will not get cancelled in a way that results in a spurious error and
thus triggers a retry.