Rework background tasks to not use shield
`aiojobs.Scheduler.shield` appears to cause a significant memory
leak for reasons that are not clear. Refactor background task
management to, instead of cancelling the tasks during shutdown, use
an `asyncio.Event` and wait on it with `asyncio.wait_for` to
implement the pauses between executions. This avoids the need for
`shield` and appears to significantly reduce the memory leak.

Refactor the way that arq workers are created in the test suite to
reuse more setup. This was an effort to eliminate a test suite memory
leak in the arq worker code, but was unsuccessful. This seems to be
specific to how the test suite is run.

Lower the thresholds for memory leak tests to account for the new
savings.