Move Qserv status to data files for testing
In the test suite, move Qserv status objects to data files and add
a new helper function to load those data files. This keeps all of
the test data together so that it can be consistently updated by
changing the data files. It also prepares for adding new data fields.