Background query start and cancel
Apparently FastStream serializes handling of all incoming Kafka
messages rather than starting a new asyncio task per message. This
caused Qserv Kafka to fail to keep up with 100 simultaneous wide
queries (with Kafka messages that were about 500KB each).

Hand off execution of query start and cancel messages to a background
asyncio task as soon as possible to unblock the Kafka consumer in
the hope that this will make the bridge more efficient with this sort
of workload.