Use batching rather than backgrounding
Rather than backgrounding query submission and immediately
acknowledging the Kafka message, which can lead to query loss if the
Qserv Kafka bridge unexpectedly restarts, instead enable batching
in the Kafka subscriber and start all the jobs in the batch in
parallel up to a configurable batch size limit. Hopefully this will
similarly improve parallelism without the loss of durability.