Add more tuning parameters and adjust settings
Add a setting for the maximum amount of data the Kafka consumer
will consume at a time and set the default so that we can consume
a full batch of wide queries. Change the default batch concurrency
to 10 from 20 because this seems to work more smoothly with Qserv.

Add a setting for the Redis connection pool size and increase the
default to account for the batch size plus some overhead for the
monitor job.

Change the default for worker jobs per result worker to 2, matching
the Phalanx configuration, and note that increasing it is probably
pointless.

Lower the Qserv database connection pool size to something more
realistic given the level of concurrency the bridge uses.