Suppress errors canceling finished jobs
If canceling a job fails in Qserv, retrieve the job status and then
quietly ignore the error if that succeeds and the job is not
executing, allowing the status to be picked up by the normal job
monitor. This handles a race condition where a result worker picks
up and processes the job just as we're trying to cancel it.