Include Kafka queue delay in metrics event
Grab the Kafka timestamp of the message in the dependency, calculate
the time spent waiting in Kafka in the handler, and propagate that
through query processing so that it can be logged in the metrics
event for a successful query. Also include the Kafka delay in the
total elapsed time for the query.