Use service discovery for DataLink rewriting
Look up the URL of the DataLink links service as an internal service
(since we don't know the dataset) and rewrite URLs flagged for
rewriting to use that URL while preserving the query string, since
the URL rewriting had only ever been used for DataLink URLs.

Refactor the VOTable encoding to use an abstract base class to hold
shared code and define a common interface, and implement that in the
BINARY2 and VOParquet encoders.